fantasense_ai_service/requirements.txt

fastapi
uvicorn
pydantic
pymongo
joblib
python-dotenv
